Refactor algos to use tfp.dist
