+++ src/garage/np/algos/cem.py
@@ -60,6 +60,12 @@ class CEM(BatchPolopt):
         self.all_params = None
         self.n_best = None
         self.n_params = None
+        self._initialize()
+
+    def _initialize(self):
+        input_var = self.env_spec.observation_space.to_tf_placeholder(
+            name='obs', batch_dims=2)
+        self.policy.build(input_var)
 
     def _sample_params(self, epoch):
         """Return sample parameters.
@@ -149,3 +155,7 @@ class CEM(BatchPolopt):
 
         logger.log(tabular)
         return rtn
+
+    def __setstate__(self, state):
+        self.__dict__ = state
+        self._initialize()
