+++ src/garage/tf/algos/__init__.py
@@ -1,6 +1,5 @@
 """Tensorflow implementation of reinforcement learning algorithms."""
 from garage.tf.algos.batch_polopt import BatchPolopt
-from garage.tf.algos.batch_polopt2 import BatchPolopt2
 from garage.tf.algos.ddpg import DDPG
 from garage.tf.algos.dqn import DQN
 from garage.tf.algos.erwr import ERWR
@@ -17,7 +16,6 @@ from garage.tf.algos.vpg import VPG
 
 __all__ = [
     'BatchPolopt',
-    'BatchPolopt2',
     'DDPG',
     'DQN',
     'ERWR',
