+++ src/garage/tf/policies/categorical_cnn_policy.py
@@ -1,15 +1,18 @@
-"""CategoricalCNNPolicy with model."""
+"""Categorical CNN Policy.
+
+A policy represented by a Categorical distribution
+which is parameterized by a convolutional neural network (CNN)
+followed a multilayer perceptron (MLP).
+"""
 import akro
+import numpy as np
 import tensorflow as tf
 
-from garage.tf.distributions import Categorical
-from garage.tf.models import CNNModel
-from garage.tf.models import MLPModel
-from garage.tf.models import Sequential
-from garage.tf.policies.policy import StochasticPolicy
+from garage.tf.models import CategoricalCNNModel
+from garage.tf.policies.policy import StochasticPolicy2
 
 
-class CategoricalCNNPolicy(StochasticPolicy):
+class CategoricalCNNPolicy(StochasticPolicy2):
     """CategoricalCNNPolicy.
 
     A policy that contains a CNN and a MLP to make prediction based on
@@ -19,18 +22,18 @@ class CategoricalCNNPolicy(StochasticPolicy):
 
     Args:
         env_spec (garage.envs.env_spec.EnvSpec): Environment specification.
-        conv_filter_sizes(tuple[int]): Dimension of the filters. For example,
+        filter_dims (tuple[int]): Dimension of the filters. For example,
             (3, 5) means there are two convolutional layers. The filter for
             first layer is of dimension (3 x 3) and the second one is of
             dimension (5 x 5).
-        conv_filters(tuple[int]): Number of filters. For example, (3, 32) means
+        num_filters (tuple[int]): Number of filters. For example, (3, 32) means
             there are two convolutional layers. The filter for the first layer
             has 3 channels and the second one with 32 channels.
-        conv_strides(tuple[int]): The stride of the sliding window. For
+        strides (tuple[int]): The stride of the sliding window. For
             example, (1, 2) means there are two convolutional layers. The
             stride of the filter for first layer is 1 and that of the second
             layer is 2.
-        conv_pad (str): The type of padding algorithm to use,
+        padding (str): The type of padding algorithm to use,
             either 'SAME' or 'VALID'.
         name (str): Policy name, also the variable scope of the policy.
         hidden_sizes (list[int]): Output dimension of dense layer(s).
@@ -60,76 +63,85 @@ class CategoricalCNNPolicy(StochasticPolicy):
 
     def __init__(self,
                  env_spec,
-                 conv_filters,
-                 conv_filter_sizes,
-                 conv_strides,
-                 conv_pad,
+                 filter_dims,
+                 num_filters,
+                 strides,
+                 padding,
                  name='CategoricalCNNPolicy',
                  hidden_sizes=(32, 32),
                  hidden_nonlinearity=tf.nn.relu,
                  hidden_w_init=tf.initializers.glorot_uniform(),
                  hidden_b_init=tf.zeros_initializer(),
-                 output_nonlinearity=tf.nn.softmax,
+                 output_nonlinearity=None,
                  output_w_init=tf.initializers.glorot_uniform(),
                  output_b_init=tf.zeros_initializer(),
                  layer_normalization=False):
-        if not isinstance(env_spec.action_space, akro.Discrete):
-            raise ValueError(
-                'CategoricalCNNPolicy only works with akro.Discrete action '
-                'space.')
-
-        if not isinstance(env_spec.observation_space, akro.Box) or \
-                not len(env_spec.observation_space.shape) in (2, 3):
-            raise ValueError(
-                '{} can only process 2D, 3D akro.Image or'
-                ' akro.Box observations, but received an env_spec with '
-                'observation_space of type {} and shape {}'.format(
-                    type(self).__name__,
-                    type(env_spec.observation_space).__name__,
-                    env_spec.observation_space.shape))
-
+        assert isinstance(env_spec.action_space, akro.Discrete), (
+            'CategoricalCNNPolicy only works with akro.Discrete action '
+            'space.')
         super().__init__(name, env_spec)
-        self.obs_dim = env_spec.observation_space.shape
-        self.action_dim = env_spec.action_space.n
+        self._obs_dim = env_spec.observation_space.shape
+        self._action_dim = env_spec.action_space.n
+        self._filter_dims = filter_dims
+        self._num_filters = num_filters
+        self._strides = strides
+        self._padding = padding
+        self._hidden_sizes = hidden_sizes
+        self._hidden_nonlinearity = hidden_nonlinearity
+        self._hidden_w_init = hidden_w_init
+        self._hidden_b_init = hidden_b_init
+        self._output_nonlinearity = output_nonlinearity
+        self._output_w_init = output_w_init
+        self._output_b_init = output_b_init
+        self._layer_normalization = layer_normalization
 
-        self.model = Sequential(
-            CNNModel(filter_dims=conv_filter_sizes,
-                     num_filters=conv_filters,
-                     strides=conv_strides,
-                     padding=conv_pad,
-                     hidden_nonlinearity=hidden_nonlinearity,
-                     name='CNNModel'),
-            MLPModel(output_dim=self.action_dim,
-                     hidden_sizes=hidden_sizes,
-                     hidden_nonlinearity=hidden_nonlinearity,
-                     hidden_w_init=hidden_w_init,
-                     hidden_b_init=hidden_b_init,
-                     output_nonlinearity=output_nonlinearity,
-                     output_w_init=output_w_init,
-                     output_b_init=output_b_init,
-                     layer_normalization=layer_normalization,
-                     name='MLPModel'))
+        self._f_prob = None
+        self._dist = None
 
-        self._initialize()
+        self.model = CategoricalCNNModel(
+            output_dim=self._action_dim,
+            filter_dims=filter_dims,
+            num_filters=num_filters,
+            strides=strides,
+            padding=padding,
+            hidden_sizes=hidden_sizes,
+            hidden_nonlinearity=hidden_nonlinearity,
+            hidden_w_init=hidden_w_init,
+            hidden_b_init=hidden_b_init,
+            output_nonlinearity=output_nonlinearity,
+            output_w_init=output_w_init,
+            output_b_init=output_b_init,
+            layer_normalization=layer_normalization)
 
-    def _initialize(self):
-        if isinstance(self.env_spec.observation_space, akro.Image):
-            state_input = tf.compat.v1.placeholder(tf.uint8,
-                                                   shape=(None, ) +
-                                                   self.obs_dim)
-            state_input = tf.cast(state_input, tf.float32)
-            state_input /= 255.0
-        else:
-            state_input = tf.compat.v1.placeholder(tf.float32,
-                                                   shape=(None, ) +
-                                                   self.obs_dim)
+    def build(self, state_input, name=None):
+        """Build model.
 
+        Args:
+          state_input (tf.Tensor) : State input.
+          name (str): Name of the model, which is also the name scope.
+
+        """
         with tf.compat.v1.variable_scope(self.name) as vs:
             self._variable_scope = vs
-            self.model.build(state_input)
+            if isinstance(self.env_spec.observation_space, akro.Image):
+                augmented_state_input = tf.cast(state_input, tf.float32)
+                augmented_state_input /= 255.0
+            else:
+                augmented_state_input = state_input
+            self._dist = self.model.build(augmented_state_input, name=name)
+            self._f_prob = tf.compat.v1.get_default_session().make_callable(
+                [tf.argmax(self._dist.sample(), -1), self._dist.probs],
+                feed_list=[state_input])
 
-        self._f_prob = tf.compat.v1.get_default_session().make_callable(
-            self.model.outputs, feed_list=[self.model.input])
+    @property
+    def distribution(self):
+        """Policy distribution.
+
+        Returns:
+            tfp.Distribution.OneHotCategorical: Policy distribution.
+
+        """
+        return self._dist
 
     @property
     def vectorized(self):
@@ -141,87 +153,63 @@ class CategoricalCNNPolicy(StochasticPolicy):
         """
         return True
 
-    def dist_info_sym(self, obs_var, state_info_vars=None, name=None):
-        """Build a symbolic graph of the distribution parameters.
-
-        Args:
-            obs_var (tf.Tensor): Tensor input for symbolic graph.
-            state_info_vars (dict[np.ndarray]): Extra state information, e.g.
-                previous action.
-            name (str): Name for symbolic graph.
-
-        Returns:
-            dict[tf.Tensor]: Outputs of the symbolic graph of distribution
-                parameters.
-
-        """
-        with tf.compat.v1.variable_scope(self._variable_scope):
-            if isinstance(self.env_spec.observation_space, akro.Image):
-                obs_var = tf.cast(obs_var, tf.float32) / 255.0
-
-            prob = self.model.build(obs_var, name=name)
-        return dict(prob=prob)
-
-    def dist_info(self, obs, state_infos=None):
-        """Get distribution parameters.
-
-        Args:
-            obs (np.ndarray): Observation input.
-            state_infos (dict[np.ndarray]): Extra state information, e.g.
-                previous action.
-
-        Returns:
-            dict[np.ndarray]: Distribution parameters.
-
-        """
-        prob = self._f_prob(obs)
-        return dict(prob=prob)
-
     def get_action(self, observation):
-        """Get single action from this policy for the input observation.
+        """Return a single action.
 
         Args:
-            observation (numpy.ndarray): Observation from environment.
+            observation (numpy.ndarray): Observations.
 
         Returns:
-            numpy.ndarray: Predicted action.
-            dict[str: np.ndarray]: Action distribution.
+            int: Action given input observation.
+            dict(numpy.ndarray): Distribution parameters.
 
         """
-        if len(observation.shape) < len(self.obs_dim):
-            observation = self.env_spec.observation_space.unflatten(
-                observation)
-        prob = self._f_prob([observation])[0]
-        action = self.action_space.weighted_sample(prob)
-        return action, dict(prob=prob)
+        sample, prob = self._f_prob(np.expand_dims([observation], 1))
+        return np.squeeze(sample), dict(prob=np.squeeze(prob, axis=1)[0])
 
     def get_actions(self, observations):
-        """Get multiple actions from this policy for the input observations.
+        """Return multiple actions.
 
         Args:
-            observations (numpy.ndarray): Observations from environment.
+            observations (numpy.ndarray): Observations.
 
         Returns:
-            numpy.ndarray: Predicted actions.
-            dict[str: np.ndarray]: Action distributions.
+            list[int]: Actions given input observations.
+            dict(numpy.ndarray): Distribution parameters.
 
         """
-        if len(observations[0].shape) < len(self.obs_dim):
-            observations = self.env_spec.observation_space.unflatten_n(
-                observations)
-        probs = self._f_prob(observations)
-        actions = list(map(self.action_space.weighted_sample, probs))
-        return actions, dict(prob=probs)
+        samples, probs = self._f_prob(np.expand_dims(observations, 1))
+        return np.squeeze(samples), dict(prob=np.squeeze(probs, axis=1))
 
-    @property
-    def distribution(self):
-        """Policy distribution.
+    def clone(self, name):
+        """Return a clone of the policy.
+
+        It only copies the configuration of the primitive,
+        not the parameters.
+
+        Args:
+            name (str): Name of the newly created policy. It has to be
+                different from source policy if cloned under the same
+                computational graph.
 
         Returns:
-            garage.tf.distributions.Categorical: Policy distribution.
+            garage.tf.policies.CategoricalCNNPolicy: Newly cloned policy.
 
         """
-        return Categorical(self.action_dim)
+        return self.__class__(name=name,
+                              env_spec=self._env_spec,
+                              filter_dims=self._filter_dims,
+                              num_filters=self._num_filters,
+                              strides=self._strides,
+                              padding=self._padding,
+                              hidden_sizes=self._hidden_sizes,
+                              hidden_nonlinearity=self._hidden_nonlinearity,
+                              hidden_w_init=self._hidden_w_init,
+                              hidden_b_init=self._hidden_b_init,
+                              output_nonlinearity=self._output_nonlinearity,
+                              output_w_init=self._output_w_init,
+                              output_b_init=self._output_b_init,
+                              layer_normalization=self._layer_normalization)
 
     def __getstate__(self):
         """Object.__getstate__.
@@ -232,14 +220,5 @@ class CategoricalCNNPolicy(StochasticPolicy):
         """
         new_dict = super().__getstate__()
         del new_dict['_f_prob']
+        del new_dict['_dist']
         return new_dict
-
-    def __setstate__(self, state):
-        """Object.__setstate__.
-
-        Args:
-            state (dict): Unpickled state.
-
-        """
-        super().__setstate__(state)
-        self._initialize()
