+++ src/garage/tf/policies/categorical_lstm_policy.py
@@ -1,18 +1,21 @@
-"""CategoricalLSTMPolicy with model."""
+"""Categorical LSTM Policy.
+
+A policy represented by a Categorical distribution
+which is parameterized by a Long short-term memory (LSTM).
+"""
 import akro
 import numpy as np
 import tensorflow as tf
 
-from garage.tf.distributions import RecurrentCategorical
-from garage.tf.models import LSTMModel
-from garage.tf.policies.policy import StochasticPolicy
+from garage.tf.models import CategoricalLSTMModel
+from garage.tf.policies.policy import StochasticPolicy2
 
 
-class CategoricalLSTMPolicy(StochasticPolicy):
-    """Estimate action distribution with Categorical parameterized by a LSTM.
+class CategoricalLSTMPolicy(StochasticPolicy2):
+    """Categorical LSTM Policy.
 
-    A policy that contains a LSTM to make prediction based on
-    a categorical distribution.
+    A policy represented by a Categorical distribution
+    which is parameterized by a Long short-term memory (LSTM).
 
     It only works with akro.Discrete action space.
 
@@ -71,7 +74,7 @@ class CategoricalLSTMPolicy(StochasticPolicy):
                  hidden_b_init=tf.zeros_initializer(),
                  recurrent_nonlinearity=tf.nn.sigmoid,
                  recurrent_w_init=tf.initializers.glorot_uniform(),
-                 output_nonlinearity=tf.nn.softmax,
+                 output_nonlinearity=None,
                  output_w_init=tf.initializers.glorot_uniform(),
                  output_b_init=tf.zeros_initializer(),
                  hidden_state_init=tf.zeros_initializer(),
@@ -88,29 +91,32 @@ class CategoricalLSTMPolicy(StochasticPolicy):
         super().__init__(name, env_spec)
         self._obs_dim = env_spec.observation_space.flat_dim
         self._action_dim = env_spec.action_space.n
+
         self._hidden_dim = hidden_dim
         self._hidden_nonlinearity = hidden_nonlinearity
         self._hidden_w_init = hidden_w_init
         self._hidden_b_init = hidden_b_init
         self._recurrent_nonlinearity = recurrent_nonlinearity
         self._recurrent_w_init = recurrent_w_init
-        self._state_include_action = state_include_action
         self._output_nonlinearity = output_nonlinearity
         self._output_w_init = output_w_init
         self._output_b_init = output_b_init
         self._hidden_state_init = hidden_state_init
         self._hidden_state_init_trainable = hidden_state_init_trainable
         self._cell_state_init = cell_state_init
-        self._cell_stat_init_trainable = cell_state_init_trainable
+        self._cell_state_init_trainable = cell_state_init_trainable
         self._forget_bias = forget_bias
         self._layer_normalization = layer_normalization
+        self._state_include_action = state_include_action
 
         if state_include_action:
             self._input_dim = self._obs_dim + self._action_dim
         else:
             self._input_dim = self._obs_dim
 
-        self.model = LSTMModel(
+        self._f_step_prob = None
+
+        self.model = CategoricalLSTMModel(
             output_dim=self._action_dim,
             hidden_dim=self._hidden_dim,
             name='prob_network',
@@ -132,29 +138,35 @@ class CategoricalLSTMPolicy(StochasticPolicy):
         self._prev_actions = None
         self._prev_hiddens = None
         self._prev_cells = None
-        self._initialize()
 
-    def _initialize(self):
-        """Initialize model."""
-        obs_ph = tf.compat.v1.placeholder(tf.float32,
-                                          shape=(None, None, self._input_dim))
-        step_input_var = tf.compat.v1.placeholder(shape=(None,
-                                                         self._input_dim),
-                                                  name='step_input',
-                                                  dtype=tf.float32)
-        step_hidden_var = tf.compat.v1.placeholder(shape=(None,
-                                                          self._hidden_dim),
-                                                   name='step_hidden_input',
-                                                   dtype=tf.float32)
-        step_cell_var = tf.compat.v1.placeholder(shape=(None,
-                                                        self._hidden_dim),
-                                                 name='step_cell_input',
-                                                 dtype=tf.float32)
+    def build(self, state_input, name=None):
+        """Build model.
 
+        Args:
+          state_input (tf.Tensor) : State input.
+          name (str): Name of the model, which is also the name scope.
+
+        """
         with tf.compat.v1.variable_scope(self.name) as vs:
             self._variable_scope = vs
-            self.model.build(obs_ph, step_input_var, step_hidden_var,
-                             step_cell_var)
+            step_input_var = tf.compat.v1.placeholder(shape=(None,
+                                                             self._input_dim),
+                                                      name='step_input',
+                                                      dtype=tf.float32)
+            step_hidden_var = tf.compat.v1.placeholder(
+                shape=(None, self._hidden_dim),
+                name='step_hidden_input',
+                dtype=tf.float32)
+            step_cell_var = tf.compat.v1.placeholder(shape=(None,
+                                                            self._hidden_dim),
+                                                     name='step_cell_input',
+                                                     dtype=tf.float32)
+
+            self.model.build(state_input,
+                             step_input_var,
+                             step_hidden_var,
+                             step_cell_var,
+                             name=name)
 
         self._f_step_prob = tf.compat.v1.get_default_session().make_callable(
             [
@@ -169,101 +181,70 @@ class CategoricalLSTMPolicy(StochasticPolicy):
         """Vectorized or not.
 
         Returns:
-            bool: True if primitive supports vectorized operations.
+            Bool: True if primitive supports vectorized operations.
 
         """
         return True
 
-    def dist_info_sym(self, obs_var, state_info_vars, name=None):
-        """Build a symbolic graph of the distribution parameters.
-
-        Args:
-            obs_var (tf.Tensor): Tensor input for symbolic graph.
-            state_info_vars (dict[np.ndarray]): Extra state information, e.g.
-                previous action.
-            name (str): Name for symbolic graph.
-
-        Returns:
-            dict[tf.Tensor]: Outputs of the symbolic graph of distribution
-                parameters.
-
-        """
-        if self._state_include_action:
-            prev_action_var = state_info_vars['prev_action']
-            prev_action_var = tf.cast(prev_action_var, tf.float32)
-            all_input_var = tf.concat(axis=2,
-                                      values=[obs_var, prev_action_var])
-        else:
-            all_input_var = obs_var
-
-        with tf.compat.v1.variable_scope(self._variable_scope):
-            outputs, _, _, _, _, _ = self.model.build(
-                all_input_var,
-                self.model.networks['default'].step_input,
-                self.model.networks['default'].step_hidden_input,
-                self.model.networks['default'].step_cell_input,
-                name=name)
-
-        return dict(prob=outputs)
-
-    def reset(self, dones=None):
+    def reset(self, do_resets=None):
         """Reset the policy.
 
         Note:
-            If `dones` is None, it will be by default `np.array([True])` which
-            implies the policy will not be "vectorized", i.e. number of
-            parallel environments for training data sampling = 1.
+            If `do_resets` is None, it will be by default np.array([True]),
+            which implies the policy will not be "vectorized", i.e. number of
+            paralle environments for training data sampling = 1.
 
         Args:
-            dones (numpy.ndarray): Bool that indicates terminal state(s).
+            do_resets (numpy.ndarray): Bool that indicates terminal state(s).
 
         """
-        if dones is None:
-            dones = [True]
-        dones = np.asarray(dones)
-        if self._prev_actions is None or len(dones) != len(self._prev_actions):
+        if do_resets is None:
+            do_resets = [True]
+        do_resets = np.asarray(do_resets)
+        if self._prev_actions is None or len(do_resets) != len(
+                self._prev_actions):
             self._prev_actions = np.zeros(
-                (len(dones), self.action_space.flat_dim))
-            self._prev_hiddens = np.zeros((len(dones), self._hidden_dim))
-            self._prev_cells = np.zeros((len(dones), self._hidden_dim))
+                (len(do_resets), self.action_space.flat_dim))
+            self._prev_hiddens = np.zeros((len(do_resets), self._hidden_dim))
+            self._prev_cells = np.zeros((len(do_resets), self._hidden_dim))
 
-        self._prev_actions[dones] = 0.
-        self._prev_hiddens[dones] = self.model.networks[
+        self._prev_actions[do_resets] = 0.
+        self._prev_hiddens[do_resets] = self.model.networks[
             'default'].init_hidden.eval()
-        self._prev_cells[dones] = self.model.networks[
+        self._prev_cells[do_resets] = self.model.networks[
             'default'].init_cell.eval()
 
     def get_action(self, observation):
-        """Get single action from this policy for the input observation.
+        """Return a single action.
 
         Args:
-            observation (numpy.ndarray): Observation from environment.
+            observation (numpy.ndarray): Observations.
 
         Returns:
-            numpy.ndarray: Predicted action.
-            dict[str: np.ndarray]: Action distribution.
+            int: Action given input observation.
+            dict(numpy.ndarray): Distribution parameters.
 
         """
         actions, agent_infos = self.get_actions([observation])
         return actions[0], {k: v[0] for k, v in agent_infos.items()}
 
     def get_actions(self, observations):
-        """Get multiple actions from this policy for the input observations.
+        """Return multiple actions.
 
         Args:
-            observations (numpy.ndarray): Observations from environment.
+            observations (numpy.ndarray): Observations.
 
         Returns:
-            numpy.ndarray: Predicted actions.
-            dict[str: np.ndarray]: Action distributions.
+            list[int]: Actions given input observations.
+            dict(numpy.ndarray): Distribution parameters.
 
         """
-        flat_obs = self.observation_space.flatten_n(observations)
         if self._state_include_action:
             assert self._prev_actions is not None
-            all_input = np.concatenate([flat_obs, self._prev_actions], axis=-1)
+            all_input = np.concatenate([observations, self._prev_actions],
+                                       axis=-1)
         else:
-            all_input = flat_obs
+            all_input = observations
         probs, hidden_vec, cell_vec = self._f_step_prob(
             all_input, self._prev_hiddens, self._prev_cells)
 
@@ -272,58 +253,49 @@ class CategoricalLSTMPolicy(StochasticPolicy):
         self._prev_actions = self.action_space.flatten_n(actions)
         self._prev_hiddens = hidden_vec
         self._prev_cells = cell_vec
-
         agent_info = dict(prob=probs)
         if self._state_include_action:
             agent_info['prev_action'] = np.copy(prev_actions)
         return actions, agent_info
 
     @property
-    def recurrent(self):
-        """Recurrent or not.
-
-        Returns:
-            bool: True if policy is recurrent.
-
-        """
-        return True
-
-    @property
     def distribution(self):
         """Policy distribution.
 
         Returns:
-            garage.tf.distributions.DiagonalGaussian: Policy distribution.
+            tfp.Distribution.OneHotCategorical: Policy distribution.
 
         """
-        return RecurrentCategorical(self._action_dim)
+        return self.model.networks['default'].dist
 
     @property
     def state_info_specs(self):
-        """State info specification.
+        """State info specifcation.
 
         Returns:
-            list[tuple]: State info specification.
+            List[str]: keys and shapes for the information related to the
+                policy's state when taking an action.
 
         """
         if self._state_include_action:
             return [
                 ('prev_action', (self._action_dim, )),
             ]
-        else:
-            return []
+        return []
 
     def clone(self, name):
         """Return a clone of the policy.
 
-        It only copies the configuration of the Q-function,
+        It only copies the configuration of the primitive,
         not the parameters.
 
         Args:
-            name (str): Name of the newly created policy.
+            name (str): Name of the newly created policy. It has to be
+                different from source policy if cloned under the same
+                computational graph.
 
         Returns:
-            garage.tf.policies.CategoricalLSTMPolicy: Clone of this object
+            garage.tf.policies.CategoricalLSTMPolicy: Newly cloned policy.
 
         """
         return self.__class__(
@@ -341,7 +313,7 @@ class CategoricalLSTMPolicy(StochasticPolicy):
             hidden_state_init=self._hidden_state_init,
             hidden_state_init_trainable=self._hidden_state_init_trainable,
             cell_state_init=self._cell_state_init,
-            cell_state_init_trainable=self._cell_stat_init_trainable,
+            cell_state_init_trainable=self._cell_state_init_trainable,
             state_include_action=self._state_include_action,
             forget_bias=self._forget_bias,
             layer_normalization=self._layer_normalization)
@@ -356,13 +328,3 @@ class CategoricalLSTMPolicy(StochasticPolicy):
         new_dict = super().__getstate__()
         del new_dict['_f_step_prob']
         return new_dict
-
-    def __setstate__(self, state):
-        """Object.__setstate__.
-
-        Args:
-            state (dict): Unpickled state.
-
-        """
-        super().__setstate__(state)
-        self._initialize()
