+++ src/garage/tf/policies/gaussian_gru_policy.py
@@ -1,14 +1,21 @@
-"""GaussianGRUPolicy with GaussianGRUModel."""
+"""Gaussian GRU Policy.
+
+A policy represented by a Gaussian distribution
+which is parameterized by a Gated Recurrent Unit (GRU).
+"""
 import akro
 import numpy as np
 import tensorflow as tf
 
-from garage.tf.models import GaussianGRUModel
-from garage.tf.policies.policy import StochasticPolicy
+from garage.tf.models import GaussianGRUModel2
+from garage.tf.policies.policy import StochasticPolicy2
 
 
-class GaussianGRUPolicy(StochasticPolicy):
-    """Models the action distribution using a Gaussian parameterized by a GRU.
+class GaussianGRUPolicy(StochasticPolicy2):
+    """Gaussian GRU Policy.
+
+    A policy represented by a Gaussian distribution
+    which is parameterized by a Gated Recurrent Unit (GRU).
 
     Args:
         env_spec (garage.envs.env_spec.EnvSpec): Environment specification.
@@ -79,7 +86,22 @@ class GaussianGRUPolicy(StochasticPolicy):
         super().__init__(name, env_spec)
         self._obs_dim = env_spec.observation_space.flat_dim
         self._action_dim = env_spec.action_space.flat_dim
+
         self._hidden_dim = hidden_dim
+        self._hidden_nonlinearity = hidden_nonlinearity
+        self._hidden_w_init = hidden_w_init
+        self._hidden_b_init = hidden_b_init
+        self._recurrent_nonlinearity = recurrent_nonlinearity
+        self._recurrent_w_init = recurrent_w_init
+        self._output_nonlinearity = output_nonlinearity
+        self._output_w_init = output_w_init
+        self._output_b_init = output_b_init
+        self._hidden_state_init = hidden_state_init
+        self._hidden_state_init_trainable = hidden_state_init_trainable
+        self._learn_std = learn_std
+        self._std_share_network = std_share_network
+        self._init_std = init_std
+        self._layer_normalization = layer_normalization
         self._state_include_action = state_include_action
 
         if state_include_action:
@@ -87,7 +109,9 @@ class GaussianGRUPolicy(StochasticPolicy):
         else:
             self._input_dim = self._obs_dim
 
-        self.model = GaussianGRUModel(
+        self._f_step_mean_std = None
+
+        self.model = GaussianGRUModel2(
             output_dim=self._action_dim,
             hidden_dim=hidden_dim,
             name='GaussianGRUModel',
@@ -108,23 +132,29 @@ class GaussianGRUPolicy(StochasticPolicy):
 
         self._prev_actions = None
         self._prev_hiddens = None
-        self._initialize()
 
-    def _initialize(self):
-        obs_ph = tf.compat.v1.placeholder(tf.float32,
-                                          shape=(None, None, self._input_dim))
-        step_input_var = tf.compat.v1.placeholder(shape=(None,
-                                                         self._input_dim),
-                                                  name='step_input',
-                                                  dtype=tf.float32)
-        step_hidden_var = tf.compat.v1.placeholder(shape=(None,
-                                                          self._hidden_dim),
-                                                   name='step_hidden_input',
-                                                   dtype=tf.float32)
+    def build(self, state_input, name=None):
+        """Build model.
 
+        Args:
+          state_input (tf.Tensor): State input.
+          name (str): Name of the model, which is also the name scope.
+
+        """
         with tf.compat.v1.variable_scope(self.name) as vs:
             self._variable_scope = vs
-            self.model.build(obs_ph, step_input_var, step_hidden_var)
+            step_input_var = tf.compat.v1.placeholder(shape=(None,
+                                                             self._input_dim),
+                                                      name='step_input',
+                                                      dtype=tf.float32)
+            step_hidden_var = tf.compat.v1.placeholder(
+                shape=(None, self._hidden_dim),
+                name='step_hidden_input',
+                dtype=tf.float32)
+            self.model.build(state_input,
+                             step_input_var,
+                             step_hidden_var,
+                             name=name)
 
         self._f_step_mean_std = (
             tf.compat.v1.get_default_session().make_callable(
@@ -137,79 +167,55 @@ class GaussianGRUPolicy(StochasticPolicy):
 
     @property
     def vectorized(self):
-        """bool: Whether the policy is vectorized or not."""
-        return True
-
-    def dist_info_sym(self, obs_var, state_info_vars, name=None):
-        """Build a symbolic graph of the distribution parameters.
-
-        Args:
-            obs_var (tf.Tensor): Tensor input for symbolic graph.
-            state_info_vars (dict): Extra state information, e.g.
-                previous action.
-            name (str): Name for symbolic graph.
+        """Vectorized or not.
 
         Returns:
-            dict[tf.Tensor]: Outputs of the symbolic graph of distribution
-                parameters.
+            Bool: True if primitive supports vectorized operations.
 
         """
-        if self._state_include_action:
-            prev_action_var = state_info_vars['prev_action']
-            prev_action_var = tf.cast(prev_action_var, tf.float32)
-            all_input_var = tf.concat(axis=2,
-                                      values=[obs_var, prev_action_var])
-        else:
-            all_input_var = obs_var
+        return True
 
-        with tf.compat.v1.variable_scope(self._variable_scope):
-            mean_var, _, log_std_var, _, _, _, _ = self.model.build(
-                all_input_var,
-                self.model.networks['default'].step_input,
-                self.model.networks['default'].step_hidden_input,
-                name=name)
-
-        return dict(mean=mean_var, log_std=log_std_var)
-
-    def reset(self, dones=None):
+    def reset(self, do_resets=None):
         """Reset the policy.
 
         Note:
-            If `dones` is None, it will be by default `np.array([True])` which
-            implies the policy will not be "vectorized", i.e. number of
+            If `do_resets` is None, it will be by default `np.array([True])`
+            which implies the policy will not be "vectorized", i.e. number of
             parallel environments for training data sampling = 1.
 
         Args:
-            dones (numpy.ndarray): Bool that indicates terminal state(s).
+            do_resets (numpy.ndarray): Bool that indicates terminal state(s).
 
         """
-        if dones is None:
-            dones = np.array([True])
-        if self._prev_actions is None or len(dones) != len(self._prev_actions):
+        if do_resets is None:
+            do_resets = np.array([True])
+        if self._prev_actions is None or len(do_resets) != len(
+                self._prev_actions):
             self._prev_actions = np.zeros(
-                (len(dones), self.action_space.flat_dim))
-            self._prev_hiddens = np.zeros((len(dones), self._hidden_dim))
+                (len(do_resets), self.action_space.flat_dim))
+            self._prev_hiddens = np.zeros((len(do_resets), self._hidden_dim))
 
-        self._prev_actions[dones] = 0.
-        self._prev_hiddens[dones] = self.model.networks[
+        self._prev_actions[do_resets] = 0.
+        self._prev_hiddens[do_resets] = self.model.networks[
             'default'].init_hidden.eval()
 
     def get_action(self, observation):
-        """Get a single action from this policy for the input observation.
+        """Get single action from this policy for the input observation.
 
         Args:
             observation (numpy.ndarray): Observation from environment.
 
         Returns:
-            tuple[numpy.ndarray, dict]: Predicted action and agent info.
+            numpy.ndarray: Actions
+            dict: Predicted action and agent information.
 
-                action (numpy.ndarray): Predicted action.
-                agent_info (dict): Distribution obtained after observing the
-                    given observation, with keys
-                    * mean: (numpy.ndarray)
-                    * log_std: (numpy.ndarray)
-                    * prev_action: (numpy.ndarray), only present if
-                        self._state_include_action is True.
+        Note:
+            It returns an action and a dict, with keys
+            - mean (numpy.ndarray): Mean of the distribution.
+            - log_std (numpy.ndarray): Log standard deviation of the
+                distribution.
+            - prev_action (numpy.ndarray): Previous action, only present if
+                self._state_include_action is True.
 
         """
         actions, agent_infos = self.get_actions([observation])
@@ -222,23 +228,24 @@ class GaussianGRUPolicy(StochasticPolicy):
             observations (numpy.ndarray): Observations from environment.
 
         Returns:
-            tuple[numpy.ndarray, dict]: Prediction actions and agent infos.
+            numpy.ndarray: Actions
+            dict: Predicted action and agent information.
 
-                actions (numpy.ndarray): Predicted actions.
-                agent_infos (dict): Distribution obtained after observing the
-                    given observation, with keys
-                    * mean: (numpy.ndarray)
-                    * log_std: (numpy.ndarray)
-                    * prev_action: (numpy.ndarray), only present if
-                        self._state_include_action is True.
+        Note:
+            It returns an action and a dict, with keys
+            - mean (numpy.ndarray): Means of the distribution.
+            - log_std (numpy.ndarray): Log standard deviations of the
+                distribution.
+            - prev_action (numpy.ndarray): Previous action, only present if
+                self._state_include_action is True.
 
         """
-        flat_obs = self.observation_space.flatten_n(observations)
         if self._state_include_action:
             assert self._prev_actions is not None
-            all_input = np.concatenate([flat_obs, self._prev_actions], axis=-1)
+            all_input = np.concatenate([observations, self._prev_actions],
+                                       axis=-1)
         else:
-            all_input = flat_obs
+            all_input = observations
         means, log_stds, hidden_vec = self._f_step_mean_std(
             all_input, self._prev_hiddens)
         rnd = np.random.normal(size=means.shape)
@@ -253,18 +260,24 @@ class GaussianGRUPolicy(StochasticPolicy):
         return samples, agent_infos
 
     @property
-    def recurrent(self):
-        """bool: Whether this policy is recurrent or not."""
-        return True
-
-    @property
     def distribution(self):
-        """garage.tf.distributions.DiagonalGaussian: Policy distribution."""
+        """Policy distribution.
+
+        Returns:
+            tfp.Distribution.MultivariateNormalDiag: Policy distribution.
+
+        """
         return self.model.networks['default'].dist
 
     @property
     def state_info_specs(self):
-        """list: State info specification."""
+        """State info specifcation.
+
+        Returns:
+            List[str]: keys and shapes for the information related to the
+                policy's state when taking an action.
+
+        """
         if self._state_include_action:
             return [
                 ('prev_action', (self._action_dim, )),
@@ -272,23 +285,48 @@ class GaussianGRUPolicy(StochasticPolicy):
 
         return []
 
-    def __getstate__(self):
-        """See `Object.__getstate__`.
+    def clone(self, name):
+        """Return a clone of the policy.
+
+        It only copies the configuration of the primitive,
+        not the parameters.
+
+        Args:
+            name (str): Name of the newly created policy. It has to be
+                different from source policy if cloned under the same
+                computational graph.
 
         Returns:
-            dict: Parameters to save.
+            garage.tf.policies.GaussianGRUPolicy: Newly cloned policy.
+
+        """
+        return self.__class__(
+            name=name,
+            env_spec=self._env_spec,
+            hidden_dim=self._hidden_dim,
+            hidden_nonlinearity=self._hidden_nonlinearity,
+            hidden_w_init=self._hidden_w_init,
+            hidden_b_init=self._hidden_b_init,
+            recurrent_nonlinearity=self._recurrent_nonlinearity,
+            recurrent_w_init=self._recurrent_w_init,
+            output_nonlinearity=self._output_nonlinearity,
+            output_w_init=self._output_w_init,
+            output_b_init=self._output_b_init,
+            hidden_state_init=self._hidden_state_init,
+            hidden_state_init_trainable=self._hidden_state_init_trainable,
+            learn_std=self._learn_std,
+            std_share_network=self._std_share_network,
+            init_std=self._init_std,
+            layer_normalization=self._layer_normalization,
+            state_include_action=self._state_include_action)
+
+    def __getstate__(self):
+        """Object.__getstate__.
+
+        Returns:
+            dict: the state to be pickled for the instance.
 
         """
         new_dict = super().__getstate__()
         del new_dict['_f_step_mean_std']
         return new_dict
-
-    def __setstate__(self, state):
-        """See `Object.__setstate__`.
-
-        Args:
-            state (dict): Parameters to restore from.
-
-        """
-        super().__setstate__(state)
-        self._initialize()
