+++ src/garage/tf/policies/gaussian_mlp_policy.py
@@ -1,17 +1,21 @@
-"""GaussianMLPPolicy with GaussianMLPModel."""
+"""Gaussian MLP Policy.
+
+A policy represented by a Gaussian distribution
+which is parameterized by a multilayer perceptron (MLP).
+"""
 import akro
 import numpy as np
 import tensorflow as tf
 
-from garage.tf.models import GaussianMLPModel
-from garage.tf.policies.policy import StochasticPolicy
+from garage.tf.models import GaussianMLPModel2
+from garage.tf.policies.policy import StochasticPolicy2
 
 
-class GaussianMLPPolicy(StochasticPolicy):
-    """GaussianMLPPolicy with GaussianMLPModel.
+class GaussianMLPPolicy(StochasticPolicy2):
+    """Gaussian MLP Policy.
 
-    A policy that contains a MLP to make prediction based on
-    a gaussian distribution.
+    A policy represented by a Gaussian distribution
+    which is parameterized by a multilayer perceptron (MLP).
 
     Args:
         env_spec (garage.envs.env_spec.EnvSpec): Environment specification.
@@ -51,16 +55,15 @@ class GaussianMLPPolicy(StochasticPolicy):
         max_std (float): If not None, the std is at most the value of max_std,
             to avoid numerical issues.
         std_hidden_nonlinearity (callable): Nonlinearity for each hidden layer
-            in the std network. It should return a tf.Tensor. Set it to None to
-            maintain a linear activation.
+            in the std network. The function should return a tf.Tensor.
         std_output_nonlinearity (callable): Nonlinearity for output layer in
-            the std network. It should return a tf.Tensor. Set it to None to
-            maintain a linear activation.
+            the std network. The function should return a
+            tf.Tensor.
         std_parameterization (str): How the std should be parametrized. There
             are a few options:
-            - exp: the logarithm of the std will be stored, and applied a
-                exponential transformation
-            - softplus: the std will be computed as log(1+exp(x))
+        - exp: the logarithm of the std will be stored, and applied a
+            exponential transformation
+        - softplus: the std will be computed as log(1+exp(x))
         layer_normalization (bool): Bool for using layer normalization or not.
 
     """
@@ -86,12 +89,37 @@ class GaussianMLPPolicy(StochasticPolicy):
                  std_output_nonlinearity=None,
                  std_parameterization='exp',
                  layer_normalization=False):
-        assert isinstance(env_spec.action_space, akro.Box)
+        if not isinstance(env_spec.action_space, akro.Box):
+            raise ValueError('GaussianMLPPolicy only works with '
+                             'akro.Box action space, but not {}'.format(
+                                 env_spec.action_space))
         super().__init__(name, env_spec)
         self.obs_dim = env_spec.observation_space.flat_dim
         self.action_dim = env_spec.action_space.flat_dim
 
-        self.model = GaussianMLPModel(
+        self._hidden_sizes = hidden_sizes
+        self._hidden_nonlinearity = hidden_nonlinearity
+        self._hidden_w_init = hidden_w_init
+        self._hidden_b_init = hidden_b_init
+        self._output_nonlinearity = output_nonlinearity
+        self._output_w_init = output_w_init
+        self._output_b_init = output_b_init
+        self._learn_std = learn_std
+        self._adaptive_std = adaptive_std
+        self._std_share_network = std_share_network
+        self._init_std = init_std
+        self._min_std = min_std
+        self._max_std = max_std
+        self._std_hidden_sizes = std_hidden_sizes
+        self._std_hidden_nonlinearity = std_hidden_nonlinearity
+        self._std_output_nonlinearity = std_output_nonlinearity
+        self._std_parameterization = std_parameterization
+        self._layer_normalization = layer_normalization
+
+        self._f_dist = None
+        self._dist = None
+
+        self.model = GaussianMLPModel2(
             output_dim=self.action_dim,
             hidden_sizes=hidden_sizes,
             hidden_nonlinearity=hidden_nonlinearity,
@@ -113,22 +141,21 @@ class GaussianMLPPolicy(StochasticPolicy):
             layer_normalization=layer_normalization,
             name='GaussianMLPModel')
 
-        self._initialize()
+    def build(self, state_input, name=None):
+        """Build model.
 
-    def _initialize(self):
-        state_input = tf.compat.v1.placeholder(tf.float32,
-                                               shape=(None, self.obs_dim))
+        Args:
+          state_input (tf.Tensor): State input.
+          name (str): Name of the model, which is also the name scope.
 
+        """
         with tf.compat.v1.variable_scope(self.name) as vs:
             self._variable_scope = vs
-            self.model.build(state_input)
-
-        self._f_dist = tf.compat.v1.get_default_session().make_callable(
-            [
-                self.model.networks['default'].mean,
-                self.model.networks['default'].log_std
-            ],
-            feed_list=[self.model.networks['default'].input])
+            self._dist = self.model.build(state_input, name=name)
+            self._f_dist = tf.compat.v1.get_default_session().make_callable(
+                [self._dist.sample(), self._dist.loc,
+                 self._dist.stddev()],
+                feed_list=[state_input])
 
     @property
     def vectorized(self):
@@ -139,24 +166,6 @@ class GaussianMLPPolicy(StochasticPolicy):
 
         """
         return True
-
-    def dist_info_sym(self, obs_var, state_info_vars=None, name='default'):
-        """Build a symbolic graph of the distribution parameters.
-
-        Args:
-            obs_var (tf.Tensor): Tensor input for symbolic graph.
-            state_info_vars (dict): Extra state information, e.g.
-                previous action.
-            name (str): Name for symbolic graph.
-
-        Returns:
-            dict[tf.Tensor]: Outputs of the symbolic graph of distribution
-                parameters.
-
-        """
-        with tf.compat.v1.variable_scope(self._variable_scope):
-            mean_var, log_std_var, _, _ = self.model.build(obs_var, name=name)
-        return dict(mean=mean_var, log_std=log_std_var)
 
     def get_action(self, observation):
         """Get single action from this policy for the input observation.
@@ -175,13 +184,10 @@ class GaussianMLPPolicy(StochasticPolicy):
                 distribution.
 
         """
-        flat_obs = self.observation_space.flatten(observation)
-        mean, log_std = self._f_dist([flat_obs])
-        rnd = np.random.normal(size=mean.shape)
-        sample = rnd * np.exp(log_std) + mean
-        sample = self.action_space.unflatten(sample[0])
-        mean = self.action_space.unflatten(mean[0])
-        log_std = self.action_space.unflatten(log_std[0])
+        sample, mean, log_std = self._f_dist(np.expand_dims([observation], 1))
+        sample = self.action_space.unflatten(np.squeeze(sample, 1)[0])
+        mean = self.action_space.unflatten(np.squeeze(mean, 1)[0])
+        log_std = self.action_space.unflatten(np.squeeze(log_std, 1)[0])
         return sample, dict(mean=mean, log_std=log_std)
 
     def get_actions(self, observations):
@@ -201,34 +207,59 @@ class GaussianMLPPolicy(StochasticPolicy):
                 distribution.
 
         """
-        flat_obs = self.observation_space.flatten_n(observations)
-        means, log_stds = self._f_dist(flat_obs)
-        rnd = np.random.normal(size=means.shape)
-        samples = rnd * np.exp(log_stds) + means
-        samples = self.action_space.unflatten_n(samples)
-        means = self.action_space.unflatten_n(means)
-        log_stds = self.action_space.unflatten_n(log_stds)
+        samples, means, log_stds = self._f_dist(np.expand_dims(
+            observations, 1))
+        samples = self.action_space.unflatten_n(np.squeeze(samples, 1))
+        means = self.action_space.unflatten_n(np.squeeze(means, 1))
+        log_stds = self.action_space.unflatten_n(np.squeeze(log_stds, 1))
         return samples, dict(mean=means, log_std=log_stds)
-
-    def get_params(self):
-        """Get the params, which are the trainable variables.
-
-        Returns:
-            List[tf.Variable]: A list of trainable variables in the current
-                variable scope.
-
-        """
-        return self.get_trainable_vars()
 
     @property
     def distribution(self):
         """Policy distribution.
 
         Returns:
-            garage.tf.distributions.DiagonalGaussian: Policy distribution.
+            tfp.Distribution.MultivariateNormalDiag: Policy distribution.
 
         """
-        return self.model.networks['default'].dist
+        return self._dist
+
+    def clone(self, name):
+        """Return a clone of the policy.
+
+        It only copies the configuration of the primitive,
+        not the parameters.
+
+        Args:
+            name (str): Name of the newly created policy. It has to be
+                different from source policy if cloned under the same
+                computational graph.
+
+        Returns:
+            garage.tf.policies.GaussianMLPPolicy: Newly cloned policy.
+
+        """
+        return self.__class__(
+            name=name,
+            env_spec=self._env_spec,
+            hidden_sizes=self._hidden_sizes,
+            hidden_nonlinearity=self._hidden_nonlinearity,
+            hidden_w_init=self._hidden_w_init,
+            hidden_b_init=self._hidden_b_init,
+            output_nonlinearity=self._output_nonlinearity,
+            output_w_init=self._output_w_init,
+            output_b_init=self._output_b_init,
+            learn_std=self._learn_std,
+            adaptive_std=self._adaptive_std,
+            std_share_network=self._std_share_network,
+            init_std=self._init_std,
+            min_std=self._min_std,
+            max_std=self._max_std,
+            std_hidden_sizes=self._std_hidden_sizes,
+            std_hidden_nonlinearity=self._std_hidden_nonlinearity,
+            std_output_nonlinearity=self._std_output_nonlinearity,
+            std_parameterization=self._std_parameterization,
+            layer_normalization=self._layer_normalization)
 
     def __getstate__(self):
         """Object.__getstate__.
@@ -239,14 +270,5 @@ class GaussianMLPPolicy(StochasticPolicy):
         """
         new_dict = super().__getstate__()
         del new_dict['_f_dist']
+        del new_dict['_dist']
         return new_dict
-
-    def __setstate__(self, state):
-        """Object.__setstate__.
-
-        Args:
-            state (dict): Unpickled state.
-
-        """
-        super().__setstate__(state)
-        self._initialize()
