+++ tests/garage/np/algos/test_cem.py
@@ -20,7 +20,6 @@ class TestCEM(TfGraphTestCase):
             policy = CategoricalMLPPolicy(name='policy',
                                           env_spec=env.spec,
                                           hidden_sizes=(32, 32))
-
             baseline = LinearFeatureBaseline(env_spec=env.spec)
 
             n_samples = 10
