+++ tests/garage/np/algos/test_cma_es.py
@@ -17,7 +17,6 @@ class TestCMAES(TfGraphTestCase):
             policy = CategoricalMLPPolicy(name='policy',
                                           env_spec=env.spec,
                                           hidden_sizes=(32, 32))
-
             baseline = LinearFeatureBaseline(env_spec=env.spec)
 
             n_samples = 20
