+++ tests/garage/tf/policies/test_categorical_lstm_policy.py
@@ -1,5 +1,4 @@
 import pickle
-from unittest import mock
 
 import numpy as np
 import pytest
@@ -10,212 +9,120 @@ from garage.tf.policies import CategoricalLSTMPolicy
 from tests.fixtures import TfGraphTestCase
 from tests.fixtures.envs.dummy import DummyBoxEnv
 from tests.fixtures.envs.dummy import DummyDiscreteEnv
-from tests.fixtures.models import SimpleLSTMModel
 
 
 class TestCategoricalLSTMPolicy(TfGraphTestCase):
 
-    @pytest.mark.parametrize('obs_dim, action_dim', [
-        ((1, ), 1),
-        ((2, ), 2),
-        ((1, 1), 1),
-        ((2, 2), 2),
-    ])
-    def test_dist_info_sym(self, obs_dim, action_dim):
-        env = TfEnv(DummyDiscreteEnv(obs_dim=obs_dim, action_dim=action_dim))
-
-        obs_ph = tf.compat.v1.placeholder(
-            tf.float32, shape=(None, None, env.observation_space.flat_dim))
-
-        with mock.patch(('garage.tf.policies.'
-                         'categorical_lstm_policy.LSTMModel'),
-                        new=SimpleLSTMModel):
-            policy = CategoricalLSTMPolicy(env_spec=env.spec,
-                                           state_include_action=False)
-
-        policy.reset()
-        obs = env.reset()
-
-        dist_sym = policy.dist_info_sym(obs_var=obs_ph,
-                                        state_info_vars=None,
-                                        name='p2_sym')
-        dist = self.sess.run(
-            dist_sym, feed_dict={obs_ph: [[obs.flatten()], [obs.flatten()]]})
-
-        assert np.array_equal(dist['prob'], np.full((2, 1, action_dim), 0.5))
-
-    @pytest.mark.parametrize('obs_dim, action_dim', [
-        ((1, ), 1),
-        ((2, ), 2),
-        ((1, 1), 1),
-        ((2, 2), 2),
-    ])
-    def test_dist_info_sym_include_action(self, obs_dim, action_dim):
-        env = TfEnv(DummyDiscreteEnv(obs_dim=obs_dim, action_dim=action_dim))
-
-        obs_ph = tf.compat.v1.placeholder(
-            tf.float32, shape=(None, None, env.observation_space.flat_dim))
-
-        with mock.patch(('garage.tf.policies.'
-                         'categorical_lstm_policy.LSTMModel'),
-                        new=SimpleLSTMModel):
-            policy = CategoricalLSTMPolicy(env_spec=env.spec,
-                                           state_include_action=True)
-
-        policy.reset()
-        obs = env.reset()
-
-        dist_sym = policy.dist_info_sym(
-            obs_var=obs_ph,
-            state_info_vars={'prev_action': np.zeros((2, 1, action_dim))},
-            name='p2_sym')
-        dist = self.sess.run(
-            dist_sym, feed_dict={obs_ph: [[obs.flatten()], [obs.flatten()]]})
-
-        assert np.array_equal(dist['prob'], np.full((2, 1, action_dim), 0.5))
-
-    def test_dist_info_sym_wrong_input(self):
-        env = TfEnv(DummyDiscreteEnv(obs_dim=(1, ), action_dim=1))
-
-        obs_ph = tf.compat.v1.placeholder(
-            tf.float32, shape=(None, None, env.observation_space.flat_dim))
-
-        with mock.patch(('garage.tf.policies.'
-                         'categorical_lstm_policy.LSTMModel'),
-                        new=SimpleLSTMModel):
-            policy = CategoricalLSTMPolicy(env_spec=env.spec,
-                                           state_include_action=True)
-
-        policy.reset()
-        obs = env.reset()
-
-        policy.dist_info_sym(
-            obs_var=obs_ph,
-            state_info_vars={'prev_action': np.zeros((3, 1, 1))},
-            name='p2_sym')
-        # observation batch size = 2 but prev_action batch size = 3
-        with pytest.raises(tf.errors.InvalidArgumentError):
-            self.sess.run(
-                policy.model.networks['p2_sym'].input,
-                feed_dict={obs_ph: [[obs.flatten()], [obs.flatten()]]})
-
     def test_invalid_env(self):
         env = TfEnv(DummyBoxEnv())
-        with mock.patch(('garage.tf.policies.'
-                         'categorical_lstm_policy.LSTMModel'),
-                        new=SimpleLSTMModel):
-            with pytest.raises(ValueError):
-                CategoricalLSTMPolicy(env_spec=env.spec)
+        with pytest.raises(ValueError):
+            CategoricalLSTMPolicy(env_spec=env.spec)
 
-    @pytest.mark.parametrize('obs_dim, action_dim', [
-        ((1, ), 1),
-        ((2, ), 2),
-        ((1, 1), 1),
-        ((2, 2), 2),
+    @pytest.mark.parametrize('obs_dim, action_dim, hidden_dim', [
+        ((1, ), 1, 4),
+        ((2, ), 2, 4),
+        ((1, 1), 1, 4),
+        ((2, 2), 2, 4),
     ])
-    @mock.patch('numpy.random.choice')
-    def test_get_action_state_include_action(self, mock_rand, obs_dim,
-                                             action_dim):
-        mock_rand.return_value = 0
-
+    def test_get_action_state_include_action(self, obs_dim, action_dim,
+                                             hidden_dim):
         env = TfEnv(DummyDiscreteEnv(obs_dim=obs_dim, action_dim=action_dim))
+        obs_var = tf.compat.v1.placeholder(
+            tf.float32,
+            shape=[None, None, env.observation_space.flat_dim + action_dim],
+            name='obs')
+        policy = CategoricalLSTMPolicy(env_spec=env.spec,
+                                       hidden_dim=hidden_dim,
+                                       state_include_action=True)
 
-        with mock.patch(('garage.tf.policies.'
-                         'categorical_lstm_policy.LSTMModel'),
-                        new=SimpleLSTMModel):
-            policy = CategoricalLSTMPolicy(env_spec=env.spec,
-                                           state_include_action=True)
-
+        policy.build(obs_var)
         policy.reset()
         obs = env.reset()
 
-        expected_prob = np.full(action_dim, 0.5)
-
-        action, agent_info = policy.get_action(obs)
+        action, _ = policy.get_action(obs.flatten())
         assert env.action_space.contains(action)
-        assert action == 0
-        assert np.array_equal(agent_info['prob'], expected_prob)
 
-        actions, agent_infos = policy.get_actions([obs])
-        for action, prob in zip(actions, agent_infos['prob']):
+        actions, _ = policy.get_actions([obs.flatten()])
+        for action in actions:
             assert env.action_space.contains(action)
-            assert action == 0
-            assert np.array_equal(prob, expected_prob)
 
-    @pytest.mark.parametrize('obs_dim, action_dim', [
-        ((1, ), 1),
-        ((2, ), 2),
-        ((1, 1), 1),
-        ((2, 2), 2),
+    @pytest.mark.parametrize('obs_dim, action_dim, hidden_dim', [
+        ((1, ), 1, 4),
+        ((2, ), 2, 4),
+        ((1, 1), 1, 4),
+        ((2, 2), 2, 4),
     ])
-    @mock.patch('numpy.random.choice')
-    def test_get_action(self, mock_rand, obs_dim, action_dim):
-        mock_rand.return_value = 0
-
+    def test_get_action(self, obs_dim, action_dim, hidden_dim):
         env = TfEnv(DummyDiscreteEnv(obs_dim=obs_dim, action_dim=action_dim))
+        obs_var = tf.compat.v1.placeholder(
+            tf.float32,
+            shape=[None, None, env.observation_space.flat_dim],
+            name='obs')
+        policy = CategoricalLSTMPolicy(env_spec=env.spec,
+                                       hidden_dim=hidden_dim,
+                                       state_include_action=False)
 
-        with mock.patch(('garage.tf.policies.'
-                         'categorical_lstm_policy.LSTMModel'),
-                        new=SimpleLSTMModel):
-            policy = CategoricalLSTMPolicy(env_spec=env.spec,
-                                           state_include_action=False)
-
+        policy.build(obs_var)
         policy.reset()
         obs = env.reset()
 
-        expected_prob = np.full(action_dim, 0.5)
-
-        action, agent_info = policy.get_action(obs)
+        action, _ = policy.get_action(obs.flatten())
         assert env.action_space.contains(action)
-        assert action == 0
-        assert np.array_equal(agent_info['prob'], expected_prob)
 
-        actions, agent_infos = policy.get_actions([obs])
-        for action, prob in zip(actions, agent_infos['prob']):
+        actions, _ = policy.get_actions([obs.flatten()])
+        for action in actions:
             assert env.action_space.contains(action)
-            assert action == 0
-            assert np.array_equal(prob, expected_prob)
 
     def test_is_pickleable(self):
         env = TfEnv(DummyDiscreteEnv(obs_dim=(1, ), action_dim=1))
-        with mock.patch(('garage.tf.policies.'
-                         'categorical_lstm_policy.LSTMModel'),
-                        new=SimpleLSTMModel):
-            policy = CategoricalLSTMPolicy(env_spec=env.spec,
-                                           state_include_action=False)
+        obs_var = tf.compat.v1.placeholder(
+            tf.float32,
+            shape=[None, None, env.observation_space.flat_dim],
+            name='obs')
+        policy = CategoricalLSTMPolicy(env_spec=env.spec,
+                                       state_include_action=False)
 
-        env.reset()
+        policy.build(obs_var)
+        policy.reset()
         obs = env.reset()
 
-        with tf.compat.v1.variable_scope('CategoricalLSTMPolicy/prob_network',
-                                         reuse=True):
-            return_var = tf.compat.v1.get_variable('return_var')
-        # assign it to all one
-        return_var.load(tf.ones_like(return_var).eval())
+        policy.model._lstm_cell.weights[0].load(
+            tf.ones_like(policy.model._lstm_cell.weights[0]).eval())
 
         output1 = self.sess.run(
-            policy.model.outputs[0],
+            [policy.distribution.probs],
             feed_dict={policy.model.input: [[obs.flatten()], [obs.flatten()]]})
 
         p = pickle.dumps(policy)
 
         with tf.compat.v1.Session(graph=tf.Graph()) as sess:
             policy_pickled = pickle.loads(p)
-            output2 = sess.run(policy_pickled.model.outputs[0],
+            obs_var = tf.compat.v1.placeholder(
+                tf.float32,
+                shape=[None, None, env.observation_space.flat_dim],
+                name='obs')
+            policy_pickled.build(obs_var)
+            output2 = sess.run([policy_pickled.distribution.probs],
                                feed_dict={
                                    policy_pickled.model.input:
                                    [[obs.flatten()], [obs.flatten()]]
                                })  # noqa: E126
             assert np.array_equal(output1, output2)
 
+    def test_state_info_specs(self):
+        env = TfEnv(DummyDiscreteEnv(obs_dim=(10, ), action_dim=4))
+        policy = CategoricalLSTMPolicy(env_spec=env.spec,
+                                       state_include_action=False)
+        assert policy.state_info_specs == []
+
+    def test_state_info_specs_with_state_include_action(self):
+        env = TfEnv(DummyDiscreteEnv(obs_dim=(10, ), action_dim=4))
+        policy = CategoricalLSTMPolicy(env_spec=env.spec,
+                                       state_include_action=True)
+        assert policy.state_info_specs == [('prev_action', (4, ))]
+
     def test_clone(self):
-        env = TfEnv(DummyDiscreteEnv(obs_dim=(1, ), action_dim=1))
-        with mock.patch(('garage.tf.policies.'
-                         'categorical_lstm_policy.LSTMModel'),
-                        new=SimpleLSTMModel):
-            policy = CategoricalLSTMPolicy(env_spec=env.spec,
-                                           state_include_action=False)
-            policy_cloned = policy.clone('cloned_policy')
-            assert policy_cloned.name == 'cloned_policy'
-            assert np.array_equal(policy.get_param_values(),
-                                  policy_cloned.get_param_values())
+        env = TfEnv(DummyDiscreteEnv(obs_dim=(10, ), action_dim=4))
+        policy = CategoricalLSTMPolicy(env_spec=env.spec)
+        policy_clone = policy.clone('CategoricalLSTMPolicyClone')
+        assert policy.env_spec == policy_clone.env_spec
