+++ tests/garage/tf/policies/test_gaussian_gru_policy.py
@@ -1,5 +1,4 @@
 import pickle
-from unittest import mock
 
 import numpy as np
 import pytest
@@ -10,99 +9,9 @@ from garage.tf.policies import GaussianGRUPolicy
 from tests.fixtures import TfGraphTestCase
 from tests.fixtures.envs.dummy import DummyBoxEnv
 from tests.fixtures.envs.dummy import DummyDiscreteEnv
-from tests.fixtures.models import SimpleGaussianGRUModel
 
 
 class TestGaussianGRUPolicy(TfGraphTestCase):
-
-    @pytest.mark.parametrize('obs_dim, action_dim, hidden_dim', [
-        ((1, ), (1, ), 4),
-        ((2, ), (2, ), 4),
-        ((1, 1), (1, ), 4),
-        ((2, 2), (2, ), 4),
-    ])
-    def test_dist_info_sym(self, obs_dim, action_dim, hidden_dim):
-        env = TfEnv(DummyBoxEnv(obs_dim=obs_dim, action_dim=action_dim))
-
-        obs_ph = tf.compat.v1.placeholder(
-            tf.float32, shape=(None, None, env.observation_space.flat_dim))
-
-        with mock.patch(('garage.tf.policies.'
-                         'gaussian_gru_policy.GaussianGRUModel'),
-                        new=SimpleGaussianGRUModel):
-            policy = GaussianGRUPolicy(env_spec=env.spec,
-                                       state_include_action=False)
-
-            policy.reset()
-            obs = env.reset()
-
-            dist_sym = policy.dist_info_sym(obs_var=obs_ph,
-                                            state_info_vars=None,
-                                            name='p2_sym')
-
-        dist = self.sess.run(
-            dist_sym, feed_dict={obs_ph: [[obs.flatten()], [obs.flatten()]]})
-
-        assert np.array_equal(dist['mean'], np.full((2, 1) + action_dim, 0.5))
-        assert np.array_equal(dist['log_std'], np.full((2, 1) + action_dim,
-                                                       0.5))
-
-    @pytest.mark.parametrize('obs_dim, action_dim, hidden_dim', [
-        ((1, ), (1, ), 4),
-        ((2, ), (2, ), 4),
-        ((1, 1), (1, ), 4),
-        ((2, 2), (2, ), 4),
-    ])
-    def test_dist_info_sym_include_action(self, obs_dim, action_dim,
-                                          hidden_dim):
-        env = TfEnv(DummyBoxEnv(obs_dim=obs_dim, action_dim=action_dim))
-
-        obs_ph = tf.compat.v1.placeholder(
-            tf.float32, shape=(None, None, env.observation_space.flat_dim))
-
-        with mock.patch(('garage.tf.policies.'
-                         'gaussian_gru_policy.GaussianGRUModel'),
-                        new=SimpleGaussianGRUModel):
-            policy = GaussianGRUPolicy(env_spec=env.spec,
-                                       state_include_action=True)
-
-            policy.reset()
-            obs = env.reset()
-            dist_sym = policy.dist_info_sym(
-                obs_var=obs_ph,
-                state_info_vars={'prev_action': np.zeros((2, 1) + action_dim)},
-                name='p2_sym')
-        dist = self.sess.run(
-            dist_sym, feed_dict={obs_ph: [[obs.flatten()], [obs.flatten()]]})
-
-        assert np.array_equal(dist['mean'], np.full((2, 1) + action_dim, 0.5))
-        assert np.array_equal(dist['log_std'], np.full((2, 1) + action_dim,
-                                                       0.5))
-
-    def test_dist_info_sym_wrong_input(self):
-        env = TfEnv(DummyBoxEnv(obs_dim=(1, ), action_dim=(1, )))
-
-        obs_ph = tf.compat.v1.placeholder(
-            tf.float32, shape=(None, None, env.observation_space.flat_dim))
-
-        with mock.patch(('garage.tf.policies.'
-                         'gaussian_gru_policy.GaussianGRUModel'),
-                        new=SimpleGaussianGRUModel):
-            policy = GaussianGRUPolicy(env_spec=env.spec,
-                                       state_include_action=True)
-
-            policy.reset()
-            obs = env.reset()
-
-            policy.dist_info_sym(
-                obs_var=obs_ph,
-                state_info_vars={'prev_action': np.zeros((3, 1, 1))},
-                name='p2_sym')
-        # observation batch size = 2 but prev_action batch size = 3
-        with pytest.raises(tf.errors.InvalidArgumentError):
-            self.sess.run(
-                policy.model.networks['p2_sym'].input,
-                feed_dict={obs_ph: [[obs.flatten()], [obs.flatten()]]})
 
     def test_invalid_env(self):
         env = TfEnv(DummyDiscreteEnv())
@@ -110,50 +19,39 @@ class TestGaussianGRUPolicy(TfGraphTestCase):
             GaussianGRUPolicy(env_spec=env.spec)
 
     # yapf: disable
-
     @pytest.mark.parametrize('obs_dim, action_dim, hidden_dim', [
         ((1, ), (1, ), 4),
         ((2, ), (2, ), 4),
         ((1, 1), (1, ), 4),
         ((2, 2), (2, ), 4)
     ])
-    @mock.patch('numpy.random.normal')
     # yapf: enable
-    def test_get_action_state_include_action(self, mock_normal, obs_dim,
-                                             action_dim, hidden_dim):
-        mock_normal.return_value = 0.5
+    def test_get_action_state_include_action(self, obs_dim, action_dim,
+                                             hidden_dim):
         env = TfEnv(DummyBoxEnv(obs_dim=obs_dim, action_dim=action_dim))
-        with mock.patch(('garage.tf.policies.'
-                         'gaussian_gru_policy.GaussianGRUModel'),
-                        new=SimpleGaussianGRUModel):
-            policy = GaussianGRUPolicy(env_spec=env.spec,
-                                       state_include_action=True)
+        obs_var = tf.compat.v1.placeholder(
+            tf.float32,
+            shape=[
+                None, None,
+                env.observation_space.flat_dim + np.prod(action_dim)
+            ],
+            name='obs')
+        policy = GaussianGRUPolicy(env_spec=env.spec,
+                                   hidden_dim=hidden_dim,
+                                   state_include_action=True)
 
+        policy.build(obs_var)
         policy.reset()
         obs = env.reset()
 
-        expected_action = np.full(action_dim, 0.5 * np.exp(0.5) + 0.5)
-        action, agent_info = policy.get_action(obs)
+        action, _ = policy.get_action(obs.flatten())
         assert env.action_space.contains(action)
-        assert np.allclose(action, expected_action)
-        expected_mean = np.full(action_dim, 0.5)
-        assert np.array_equal(agent_info['mean'], expected_mean)
-        expected_log_std = np.full(action_dim, 0.5)
-        assert np.array_equal(agent_info['log_std'], expected_log_std)
-        expected_prev_action = np.full(action_dim, 0)
-        assert np.array_equal(agent_info['prev_action'], expected_prev_action)
 
         policy.reset()
 
-        actions, agent_infos = policy.get_actions([obs])
-        for action, mean, log_std, prev_action in zip(
-                actions, agent_infos['mean'], agent_infos['log_std'],
-                agent_infos['prev_action']):
+        actions, _ = policy.get_actions([obs.flatten()])
+        for action in actions:
             assert env.action_space.contains(action)
-            assert np.allclose(action, expected_action)
-            assert np.array_equal(mean, expected_mean)
-            assert np.array_equal(log_std, expected_log_std)
-            assert np.array_equal(prev_action, expected_prev_action)
 
     # yapf: disable
     @pytest.mark.parametrize('obs_dim, action_dim, hidden_dim', [
@@ -162,69 +60,88 @@ class TestGaussianGRUPolicy(TfGraphTestCase):
         ((1, 1), (1, ), 4),
         ((2, 2), (2, ), 4)
     ])
-    @mock.patch('numpy.random.normal')
     # yapf: enable
-    def test_get_action(self, mock_normal, obs_dim, action_dim, hidden_dim):
-        mock_normal.return_value = 0.5
+    def test_get_action(self, obs_dim, action_dim, hidden_dim):
         env = TfEnv(DummyBoxEnv(obs_dim=obs_dim, action_dim=action_dim))
-        with mock.patch(('garage.tf.policies.'
-                         'gaussian_gru_policy.GaussianGRUModel'),
-                        new=SimpleGaussianGRUModel):
-            policy = GaussianGRUPolicy(env_spec=env.spec,
-                                       state_include_action=False)
+        obs_var = tf.compat.v1.placeholder(
+            tf.float32,
+            shape=[None, None, env.observation_space.flat_dim],
+            name='obs')
+        policy = GaussianGRUPolicy(env_spec=env.spec,
+                                   hidden_dim=hidden_dim,
+                                   state_include_action=False)
 
-        policy.reset()
+        policy.build(obs_var)
+        policy.reset(do_resets=None)
         obs = env.reset()
 
-        expected_action = np.full(action_dim, 0.5 * np.exp(0.5) + 0.5)
-        action, agent_info = policy.get_action(obs)
+        action, _ = policy.get_action(obs.flatten())
         assert env.action_space.contains(action)
-        assert np.allclose(action, expected_action)
 
-        expected_mean = np.full(action_dim, 0.5)
-        assert np.array_equal(agent_info['mean'], expected_mean)
-        expected_log_std = np.full(action_dim, 0.5)
-        assert np.array_equal(agent_info['log_std'], expected_log_std)
-
-        actions, agent_infos = policy.get_actions([obs])
-        for action, mean, log_std in zip(actions, agent_infos['mean'],
-                                         agent_infos['log_std']):
+        actions, _ = policy.get_actions([obs.flatten()])
+        for action in actions:
             assert env.action_space.contains(action)
-            assert np.allclose(action, expected_action)
-            assert np.array_equal(mean, expected_mean)
-            assert np.array_equal(log_std, expected_log_std)
 
     def test_is_pickleable(self):
         env = TfEnv(DummyBoxEnv(obs_dim=(1, ), action_dim=(1, )))
-        with mock.patch(('garage.tf.policies.'
-                         'gaussian_gru_policy.GaussianGRUModel'),
-                        new=SimpleGaussianGRUModel):
-            policy = GaussianGRUPolicy(env_spec=env.spec,
-                                       state_include_action=False)
+        obs_var = tf.compat.v1.placeholder(
+            tf.float32,
+            shape=[None, None, env.observation_space.flat_dim],
+            name='obs')
+        policy = GaussianGRUPolicy(env_spec=env.spec,
+                                   state_include_action=False)
 
+        policy.build(obs_var)
         env.reset()
         obs = env.reset()
-
         with tf.compat.v1.variable_scope('GaussianGRUPolicy/GaussianGRUModel',
                                          reuse=True):
-            return_var = tf.compat.v1.get_variable('return_var')
+            param = tf.compat.v1.get_variable(
+                'dist_params/log_std_param/parameter')
         # assign it to all one
-        return_var.load(tf.ones_like(return_var).eval())
+        param.load(tf.ones_like(param).eval())
 
         output1 = self.sess.run(
-            policy.model.networks['default'].mean,
+            [policy.distribution.loc,
+             policy.distribution.stddev()],
             feed_dict={policy.model.input: [[obs.flatten()], [obs.flatten()]]})
 
         p = pickle.dumps(policy)
 
         with tf.compat.v1.Session(graph=tf.Graph()) as sess:
             policy_pickled = pickle.loads(p)
+            obs_var = tf.compat.v1.placeholder(
+                tf.float32,
+                shape=[None, None, env.observation_space.flat_dim],
+                name='obs')
+            policy_pickled.build(obs_var)
             # yapf: disable
             output2 = sess.run(
-                policy_pickled.model.networks['default'].mean,
+                [
+                    policy_pickled.distribution.loc,
+                    policy_pickled.distribution.stddev()
+                ],
                 feed_dict={
                     policy_pickled.model.input: [[obs.flatten()],
                                                  [obs.flatten()]]
                 })
             assert np.array_equal(output1, output2)
             # yapf: enable
+
+    def test_state_info_specs(self):
+        env = TfEnv(DummyBoxEnv(obs_dim=(4, ), action_dim=(4, )))
+        policy = GaussianGRUPolicy(env_spec=env.spec,
+                                   state_include_action=False)
+        assert policy.state_info_specs == []
+
+    def test_state_info_specs_with_state_include_action(self):
+        env = TfEnv(DummyBoxEnv(obs_dim=(4, ), action_dim=(4, )))
+        policy = GaussianGRUPolicy(env_spec=env.spec,
+                                   state_include_action=True)
+        assert policy.state_info_specs == [('prev_action', (4, ))]
+
+    def test_clone(self):
+        env = TfEnv(DummyBoxEnv(obs_dim=(4, ), action_dim=(4, )))
+        policy = GaussianGRUPolicy(env_spec=env.spec)
+        policy_clone = policy.clone('GaussianGRUPolicyClone')
+        assert policy_clone.env_spec == policy.env_spec
