+++ tests/garage/tf/policies/test_gaussian_mlp_policy.py
@@ -1,5 +1,4 @@
 import pickle
-from unittest import mock
 
 import numpy as np
 import pytest
@@ -9,11 +8,16 @@ from garage.tf.envs import TfEnv
 from garage.tf.policies import GaussianMLPPolicy
 from tests.fixtures import TfGraphTestCase
 from tests.fixtures.envs.dummy import DummyBoxEnv
-from tests.fixtures.models import SimpleGaussianMLPModel
+from tests.fixtures.envs.dummy import DummyDiscreteEnv
 
 
 class TestGaussianMLPPolicy(TfGraphTestCase):
 
+    def test_invalid_env(self):
+        env = TfEnv(DummyDiscreteEnv())
+        with pytest.raises(ValueError):
+            GaussianMLPPolicy(env_spec=env.spec)
+
     @pytest.mark.parametrize('obs_dim, action_dim', [
         ((1, ), (1, )),
         ((1, ), (2, )),
@@ -22,68 +26,25 @@ class TestGaussianMLPPolicy(TfGraphTestCase):
         ((1, 1), (2, 2)),
         ((2, 2), (2, 2)),
     ])
-    @mock.patch('numpy.random.normal')
-    def test_get_action(self, mock_normal, obs_dim, action_dim):
-        mock_normal.return_value = 0.5
+    def test_get_action(self, obs_dim, action_dim):
         env = TfEnv(DummyBoxEnv(obs_dim=obs_dim, action_dim=action_dim))
-        with mock.patch(('garage.tf.policies.'
-                         'gaussian_mlp_policy.GaussianMLPModel'),
-                        new=SimpleGaussianMLPModel):
-            policy = GaussianMLPPolicy(env_spec=env.spec)
+        obs_var = tf.compat.v1.placeholder(
+            tf.float32,
+            shape=[None, None, env.observation_space.flat_dim],
+            name='obs')
+        policy = GaussianMLPPolicy(env_spec=env.spec)
 
+        policy.build(obs_var)
         env.reset()
         obs, _, _, _ = env.step(1)
 
-        action, prob = policy.get_action(obs)
-
-        expected_action = np.full(action_dim, 0.75)
-        expected_mean = np.full(action_dim, 0.5)
-        expected_log_std = np.full(action_dim, np.log(0.5))
-
+        action, _ = policy.get_action(obs.flatten())
         assert env.action_space.contains(action)
-        assert np.array_equal(action, expected_action)
-        assert np.array_equal(prob['mean'], expected_mean)
-        assert np.array_equal(prob['log_std'], expected_log_std)
-
-        actions, probs = policy.get_actions([obs, obs, obs])
-        for action, mean, log_std in zip(actions, probs['mean'],
-                                         probs['log_std']):
+        actions, _ = policy.get_actions(
+            [obs.flatten(), obs.flatten(),
+             obs.flatten()])
+        for action in actions:
             assert env.action_space.contains(action)
-            assert np.array_equal(action, expected_action)
-            assert np.array_equal(prob['mean'], expected_mean)
-            assert np.array_equal(prob['log_std'], expected_log_std)
-
-    @pytest.mark.parametrize('obs_dim, action_dim', [
-        ((1, ), (1, )),
-        ((1, ), (2, )),
-        ((2, ), (2, )),
-        ((1, 1), (1, 1)),
-        ((1, 1), (2, 2)),
-        ((2, 2), (2, 2)),
-    ])
-    def test_dist_info_sym(self, obs_dim, action_dim):
-        env = TfEnv(DummyBoxEnv(obs_dim=obs_dim, action_dim=action_dim))
-        with mock.patch(('garage.tf.policies.'
-                         'gaussian_mlp_policy.GaussianMLPModel'),
-                        new=SimpleGaussianMLPModel):
-            policy = GaussianMLPPolicy(env_spec=env.spec)
-
-        env.reset()
-        obs, _, _, _ = env.step(1)
-
-        obs_dim = env.spec.observation_space.flat_dim
-        obs_ph = tf.compat.v1.placeholder(tf.float32, shape=(None, obs_dim))
-
-        dist1_sym = policy.dist_info_sym(obs_ph, name='p1_sym')
-
-        # flatten output
-        expected_mean = [np.full(np.prod(action_dim), 0.5)]
-        expected_log_std = [np.full(np.prod(action_dim), np.log(0.5))]
-
-        prob = self.sess.run(dist1_sym, feed_dict={obs_ph: [obs.flatten()]})
-
-        assert np.array_equal(prob['mean'], expected_mean)
-        assert np.array_equal(prob['log_std'], expected_log_std)
 
     @pytest.mark.parametrize('obs_dim, action_dim', [
         ((1, ), (1, )),
@@ -95,28 +56,44 @@ class TestGaussianMLPPolicy(TfGraphTestCase):
     ])
     def test_is_pickleable(self, obs_dim, action_dim):
         env = TfEnv(DummyBoxEnv(obs_dim=obs_dim, action_dim=action_dim))
-        with mock.patch(('garage.tf.policies.'
-                         'gaussian_mlp_policy.GaussianMLPModel'),
-                        new=SimpleGaussianMLPModel):
-            policy = GaussianMLPPolicy(env_spec=env.spec)
+        obs_var = tf.compat.v1.placeholder(
+            tf.float32,
+            shape=[None, None, env.observation_space.flat_dim],
+            name='obs')
+        policy = GaussianMLPPolicy(env_spec=env.spec)
 
-        env.reset()
-        obs, _, _, _ = env.step(1)
-        obs_dim = env.spec.observation_space.flat_dim
+        policy.build(obs_var)
+        obs = env.reset()
 
         with tf.compat.v1.variable_scope('GaussianMLPPolicy/GaussianMLPModel',
                                          reuse=True):
-            return_var = tf.compat.v1.get_variable('return_var')
+            bias = tf.compat.v1.get_variable(
+                'dist_params/mean_network/hidden_0/bias')
         # assign it to all one
-        return_var.load(tf.ones_like(return_var).eval())
+        bias.load(tf.ones_like(bias).eval())
         output1 = self.sess.run(
-            policy.model.outputs[:-1],
-            feed_dict={policy.model.input: [obs.flatten()]})
+            [policy.distribution.loc,
+             policy.distribution.stddev()],
+            feed_dict={policy.model.input: [[obs.flatten()]]})
 
         p = pickle.dumps(policy)
         with tf.compat.v1.Session(graph=tf.Graph()) as sess:
+            obs_var = tf.compat.v1.placeholder(
+                tf.float32,
+                shape=[None, None, env.observation_space.flat_dim],
+                name='obs')
             policy_pickled = pickle.loads(p)
+            policy_pickled.build(obs_var)
             output2 = sess.run(
-                policy_pickled.model.outputs[:-1],
-                feed_dict={policy_pickled.model.input: [obs.flatten()]})
+                [
+                    policy_pickled.distribution.loc,
+                    policy_pickled.distribution.stddev()
+                ],
+                feed_dict={policy_pickled.model.input: [[obs.flatten()]]})
             assert np.array_equal(output1, output2)
+
+    def test_clone(self):
+        env = TfEnv(DummyBoxEnv(obs_dim=(10, ), action_dim=(4, )))
+        policy = GaussianMLPPolicy(env_spec=env.spec)
+        policy_clone = policy.clone('GaussnaMLPPolicyClone')
+        assert policy.env_spec == policy_clone.env_spec
