+++ tests/garage/tf/policies/test_policies.py
@@ -17,26 +17,30 @@ class TestDiscretePolicies(TfGraphTestCase):
     def setup_method(self):
         super().setup_method()
         self.env = TfEnv(DummyDiscreteEnv())
+        self.obs_var = tf.compat.v1.placeholder(
+            tf.float32,
+            shape=[None, None, self.env.observation_space.flat_dim],
+            name='obs')
 
     def teardown_method(self):
         self.env.close()
         super().teardown_method()
 
     def test_categorial_gru_policy(self):
-        categorical_gru_policy = CategoricalGRUPolicy(env_spec=self.env,
-                                                      hidden_dim=1)
+        categorical_gru_policy = CategoricalGRUPolicy(
+            env_spec=self.env, hidden_dim=1, state_include_action=False)
         self.sess.run(tf.compat.v1.global_variables_initializer())
-
+        categorical_gru_policy.build(self.obs_var)
         categorical_gru_policy.reset()
 
         obs = self.env.observation_space.high
         assert categorical_gru_policy.get_action(obs)
 
     def test_categorical_lstm_policy(self):
-        categorical_lstm_policy = CategoricalLSTMPolicy(env_spec=self.env,
-                                                        hidden_dim=1)
+        categorical_lstm_policy = CategoricalLSTMPolicy(
+            env_spec=self.env, hidden_dim=1, state_include_action=False)
         self.sess.run(tf.compat.v1.global_variables_initializer())
-
+        categorical_lstm_policy.build(self.obs_var)
         categorical_lstm_policy.reset()
 
         obs = self.env.observation_space.high
@@ -46,6 +50,7 @@ class TestDiscretePolicies(TfGraphTestCase):
         categorical_mlp_policy = CategoricalMLPPolicy(env_spec=self.env,
                                                       hidden_sizes=(1, ))
         self.sess.run(tf.compat.v1.global_variables_initializer())
+        categorical_mlp_policy.build(self.obs_var)
 
         obs = self.env.observation_space.high
         assert categorical_mlp_policy.get_action(obs)
@@ -56,6 +61,10 @@ class TestContinuousPolicies(TfGraphTestCase):
     def setup_method(self):
         super().setup_method()
         self.env = TfEnv(DummyBoxEnv())
+        self.obs_var = tf.compat.v1.placeholder(
+            tf.float32,
+            shape=[None, None, self.env.observation_space.flat_dim],
+            name='obs')
 
     def teardown_method(self):
         self.env.close()
@@ -71,9 +80,11 @@ class TestContinuousPolicies(TfGraphTestCase):
 
     def test_gaussian_gru_policy(self):
         gaussian_gru_policy = GaussianGRUPolicy(env_spec=self.env,
-                                                hidden_dim=1)
+                                                hidden_dim=1,
+                                                state_include_action=False)
         self.sess.run(tf.compat.v1.global_variables_initializer())
 
+        gaussian_gru_policy.build(self.obs_var)
         gaussian_gru_policy.reset()
 
         obs = self.env.observation_space.high
@@ -81,9 +92,11 @@ class TestContinuousPolicies(TfGraphTestCase):
 
     def test_gaussian_lstm_policy(self):
         gaussian_lstm_policy = GaussianLSTMPolicy(env_spec=self.env,
-                                                  hidden_dim=1)
+                                                  hidden_dim=1,
+                                                  state_include_action=False)
         self.sess.run(tf.compat.v1.global_variables_initializer())
 
+        gaussian_lstm_policy.build(self.obs_var)
         gaussian_lstm_policy.reset()
 
         obs = self.env.observation_space.high
@@ -93,6 +106,7 @@ class TestContinuousPolicies(TfGraphTestCase):
         gaussian_mlp_policy = GaussianMLPPolicy(env_spec=self.env,
                                                 hidden_sizes=(1, ))
         self.sess.run(tf.compat.v1.global_variables_initializer())
+        gaussian_mlp_policy.build(self.obs_var)
 
         obs = self.env.observation_space.high
         assert gaussian_mlp_policy.get_action(obs)
